requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
frozenlist==1.4.1
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-core==0.2.7
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.77
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==1.34.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
